chore(ffi): move MsgLike related types to their own file
